frontend to get some view from backend

+++ src/main/java/eu/chowaniok/Ping.java
@@ -8,14 +8,14 @@ import lombok.Setter;
  */
 public class Ping {
 
-    @Getter @Setter private long id;
+    @Getter @Setter private double id;
     @Getter @Setter private int year;
     @Getter @Setter private int month;
     @Getter @Setter private int day;
     @Getter @Setter private int hour;
     @Getter @Setter private int minute;
 
-    public Ping(long id, int year, int month, int day, int hour, int minute) {
+    public Ping(double id, int year, int month, int day, int hour, int minute) {
         this.id = id;
         this.year = year;
         this.month = month;
